+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Duplicate diagnostic for member with changed surname
 - No needed space for next relation while formating
+- Bug with completion in unknown member
 
 ## [1.6.0] - 2025-01-07