+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Duplicate diagnostic for member with changed surname
 - No needed space for next relation while formating
 - Bug with completion in unknown member
+- Reference name with surname from children of regular relations
 
 ## [1.6.0] - 2025-01-07